Removed props in TabNavigator relevant to AddScreen
MapScreen was cleaned up to no longer receive addRestroom function as a prop. It is also no longer used as a dependency in the component useEffect. The map on MapScreen automatically re-renders upon addition of a new restroom.
Navigation is no longer passed in to screens from TabNavigator as navigation props is implicity passed in. Navigation must only be passed into child components who are not registered routes in RootStackNavigator or TabNavigator